requirements.txt
flask
qrcode
pillow
requests
cloudinary
pymongo
gunicorn
